Finishes CSS. Finished validation checks.

+++ src/persistence/PersistenceUtil.java
@@ -10,7 +10,7 @@ import javax.persistence.Persistence;
 import entity.ToDoObject;
 import entity.User;
 
-@SuppressWarnings("serial")
+@SuppressWarnings({"serial", "unchecked"})
 public class PersistenceUtil implements Serializable {
 	protected static EntityManagerFactory emf = Persistence.createEntityManagerFactory("DSS1");
 
@@ -31,7 +31,7 @@ public class PersistenceUtil implements Serializable {
 		em.close();
 	}
 	
-	public static User findUser(String userName){
+	public static User findSingleUserName(String userName){
 		EntityManager em = emf.createEntityManager();
 		User user = em.find(User.class, userName);
 		em.close();
@@ -43,9 +43,37 @@ public class PersistenceUtil implements Serializable {
 		EntityManager em = emf.createEntityManager();
 		ToDoObject objectToRemove = em.find(ToDoObject.class, id);
 		
-		em.getTransaction().begin();
-		em.remove(objectToRemove);
-		em.getTransaction().commit();
-		em.close();
+		if(objectToRemove != null){ //If connection is slow and user double-clicks "remove", system tried to remove object twice
+			em.getTransaction().begin();
+			em.remove(objectToRemove);
+			em.getTransaction().commit();
+			em.close();
+		}
 	}
+	
+	public static User findUserAndPasswordByUsername(String userName){
+		EntityManager em = emf.createEntityManager();
+		List<User> users = (List<User>) em.createNamedQuery("User.findUserAndPasswordByUsername").setParameter("userName", userName).getResultList();
+		em.close();
+		
+		if(users.isEmpty())
+			return null;
+		else return users.get(0);
+	}
+	
+	public static List<String> findAllUserNames(){
+		EntityManager em = emf.createEntityManager();
+		List<String> users = (List<String>) em.createNamedQuery("User.findAllUserNames").getResultList();
+		em.close();
+		
+		return users;
+	}
+	
+//	public static List<String[]> findToDoObjectsByUsername(String userName){
+//		EntityManager em = emf.createEntityManager();
+//		List<String[]> toDoObjects = (List<String[]>) em.createNamedQuery("ToDoObject.findToDoObjectsByUsername").setParameter("userName", userName).getResultList();
+//		em.close();
+//		
+//		return toDoObjects;
+//	}
 }
